feat - Add support for storing results of an expectation (#14)
## Description
This update focuses on implementing the functionality to store data from
expectations. Stored data can be "scoped" to the entire file or just the
current spec's expectations.
This update does not include support for accessing a stored value from
specs. The next update will focus on implementing it.
I also reworked `Runner` a bit to group and move logic into separate
classes, `Metadata` and `Callbacks`. This allowed me to add specs around
the lifecycle, which was nice.

## Changelog

### Added
- Added new `Runner::Callbacks` class to handle various events while
specs are running
- Added new `Runner::Metadata` class to handle setting example metadata
for error reporting
- Added support for defining stored expectations via the `store_as` key.
  ```yaml
  create_user:
    path: "/users"
    method: "post"
    expectations:
    - variables:
        name: "John"
        email: "john@example.com"
      store_as: "created_user"
      expect:
        status: 200
  ```

### Changed
- Finished out `Context::Store` 
- Updated `everythingrb`

### Removed
- Removed `Context::Metadata` as it is not needed right now