feat - Add support for referencing stored test data (#15)
## Description
This update is part 2 of the storing test data. This adds
`Attribute::Store` that can referenced stored test data via the `store`
keyword.

Both file and spec scoped entries can be accessed via this keyword. When
accessing stored data scoped to the spec, do not provide the "spec"
prefix as the are not stored with it.

```yaml
  create_user:
    path: "/users"
    method: "post"
    expectations:
    - variables:
        name: "John"
        email: "john@example.com"
      store_as: "created_user"
      expect:
        status: 200

  show_user:
    path: "/users/:id"
    query:
      id: store.created_user.response.id
    - expect:
        status: 200
  ```

## Changelog

### Added
- Added `Attribute::Store` for accessing stored test data

### Changed
- Cleaned up `Attribute.from_string`