feat - Added compound matcher `matcher.and` (#17)
## Description
This PR adds support for SpecForge's first custom matcher,
`matcher.and`. This is similar to RSpec's `and` matcher, but I felt the
debugging experience with RSpec's version made it difficult to
understand what was going on so I vastly improved the error reporting
when something does fail. While I was adding this matcher, it became
apparent to me that the resolution methods on Attribute were becoming
smelly. I went through and gave them some TLC.

## Changelog

### Added
- Added `Attribute#resolve_as_matcher` that resolves the attribute into
an RSpec matcher
- Added `matcher.and` for compound matching 
```yaml
email:
  matcher.and:
  - kind_of.string
  - /@/
  - matcher.end_with: ".com"
```
- Added custom RSpec matcher `have_size` for checking an objects size.
Accessible via `matcher.have_size`

### Changed
- Renamed `Attribute#resolve` to `#resolved`.
- Renamed `Attribute#resolve_value` to `#resolve`.
- Updated integration tests to use `matcher.and`.