feat - Add callback support (#19)
## Description
This PR adds support for users to define and reference callbacks on a
per file basis. Callbacks are defined during configuration, don't
require commitment of when they are triggered ahead of time, and are
provided a context specific object that changes depending on when the
callback is triggered.
```rb
SpecForge.configure do |config|
  config.define_callback :prepare_database do |context|
    DatabaseCleaner.start
  end
end 
```
Once a callback is defined, it can be used in a spec file under the
`global` key:
```yaml
# specs/users.yml
global:
  callbacks:
  - before: prepare_database # alias for before_each
```
Other callbacks are before_file, before_spec, before_each, after_file,
after_spec, and after_each.