chore - Updated documentation and removed overlay (#22)
## Description
This PR updates the README.md to add information about the newer
features. While I was working on it, I uncovered a bug that was caused
by a choice I made when originally authoring the
normalizer/overlay_options. If someone were to attempt to overwrite an
attribute with an empty value, let's say `""`, overlay_options would
ignore that value and use the original value. This was caused by
choosing to use empty values as defaults in the normalizer.

Will this come and bite me in the ass later like that original decision
did? Only time will tell! :laughing: